Format fix: s?size_t on Windows is long. (st)
git-svn-id: svn://mielke.cc/main/brltty@1859 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ Programs/api_client.c
@@ -195,7 +195,7 @@ readNextPacket:
     handle_keycode(ntohl(*code));
     goto readNextPacket;
   }
-  syslog(LOG_ERR,"(brlapi_waitForPacket) Received unexpected packet of type %s and size %d\n",brlapi_packetType(type),res);
+  syslog(LOG_ERR,"(brlapi_waitForPacket) Received unexpected packet of type %s and size %ld\n",brlapi_packetType(type),(long)res);
   goto readNextPacket;
 
 out:
